pyproject.toml
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[project]
name = "django-countries"
version = "8.0.1"
description = "Provides a country field for Django models."
readme = "README.md"
authors = [{ name = "Chris Beaven", email = "smileychris@gmail.com" }]
license = { text = "MIT" }
keywords = ["django", "countries", "flags"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
]
requires-python = ">=3.8"
dependencies = ["asgiref", "typing_extensions"]

[project.urls]
Homepage = "https://github.com/SmileyChris/django-countries/"
"Change Log" = "https://github.com/SmileyChris/django-countries/blob/main/CHANGES.md"
"Source Code" = "https://github.com/SmileyChris/django-countries"
Translations = "https://explore.transifex.com/smileychris/django-countries/"

[project.optional-dependencies]
pyuca = ["pyuca"]

[dependency-groups]
test = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "djangorestframework",
    "graphene-django",
    "coverage",
    "drf-spectacular",
]
dev = [
    { include-group = "test" },
    "ruff",
    "django",
    "mypy>=1.12,<1.16",
    "django-stubs",
    "djangorestframework-stubs",
    "Pillow",
    "types-Pillow",
    "bandit",
    "docutils",
    "Pygments",
    "towncrier",
]
docs = ["mkdocs", "mkdocs-material"]

[tool.uv.build-backend]
module-root = ""
source-include = ["LICENSE", "CHANGES.md"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_countries.tests.settings"
filterwarnings = ["ignore::DeprecationWarning:graphene"]

[tool.coverage.run]
source = ["django_countries"]
omit = ["django_countries/makesprite.py", "django_countries/tests/settings*.py"]
parallel = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "@overload", "if TYPE_CHECKING:"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
disallow_untyped_calls = true
python_version = "3.9"
exclude = [
    'django_countries/tests/',
]

[[tool.mypy.overrides]]
module = "mypy_django_plugin.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "django_countries.tests.settings"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
# Enable pycodestyle (E, W), Pyflakes (F), isort (I), pyupgrade (UP)
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "PIE",  # flake8-pie
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "PERF", # Performance anti-patterns
    "TCH",  # Type-checking optimizations
    "RSE",  # Raise statement improvements
    "DJ",   # Django best practices
    "T20",  # Print statement detection
]

ignore = [
    "DJ008", # Model __str__ not critical for test models
]

[tool.ruff.lint.per-file-ignores]
"django_countries/tests/**" = [
    "S101", # Allow asserts in tests
    "ARG",  # Allow unused arguments in tests
    "N802", # Allow test names matching settings
    "N806", # Allow Form class variables
]
"django_countries/data.py" = ["T201"] # Allow prints in data generation script
"django_countries/ioc_data.py" = [
    "T201",
] # Allow prints in data generation script
"django_countries/graphql/**" = [
    "ARG004",
] # GraphQL resolvers must match signature

[tool.ruff.lint.isort]
known-first-party = ["django_countries"]

[tool.bandit]
exclude_dirs = ["/tests"]

[tool.towncrier]
package = "django_countries"
filename = "CHANGES.md"
directory = "changes"
issue_format = "[#{issue}](https://github.com/SmileyChris/django-countries/issues/{issue})"
title_format = "## {version} ({project_date})"
